+++ src/camera.h
@@ -11,7 +11,7 @@ public:
     void SetPerspectiveProjection(float fov, float aspect, float near, float far);
     void SetOrthographicProjection(float left, float right, float top, float bottom, float near, float far);
 
-    void SetViewDirection(glm::vec3 direction, glm::dvec3 up = glm::dvec3{0.0f, -1.0f, 0.0f});
+    void SetViewDirection(glm::dvec3 direction, glm::dvec3 up = glm::dvec3{0.0f, -1.0f, 0.0f});
     void SetViewTarget(glm::dvec3 target, glm::dvec3 up = glm::dvec3{0.0f, -1.0f, 0.0f});
     void SetViewYXZ(glm::vec3& rotation);
